Refactor cache,security

+++ src/main/resources/db/migration/V1.5__Add_reviews_table.sql
@@ -6,7 +6,7 @@ CREATE TABLE reviews
 
 CREATE TABLE users
 (
-    "id"       BIGINT PRIMARY KEY,
+    "id"       SERIAL PRIMARY KEY,
     "username" VARCHAR(32) NOT NULL
 );
